requirements.txt
# Minimal deps for production (Render). Full old export: requirements-legacy.txt
setuptools>=69.0
wheel>=0.42.0
Django==4.2.16
asgiref==3.8.1
sqlparse==0.5.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-environ==0.11.2
django-jazzmin==2.6.0
django-import-export==4.1.1
Pillow==10.4.0
gunicorn==23.0.0
firebase-admin==6.5.0
django-cors-headers==4.3.1
